created invoice table and invice items tables

+++ src/db/models/invoice_item.model.js
@@ -30,7 +30,8 @@ export const InvoiceItem = sequelize.define(
       type: DataTypes.INTEGER,
     },
     salePrice: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DOUBLE,
+      
     },
   },
   {
